requirements.txt
requests
pdfplumber
sentence-transformers
faiss-cpu
rank-bm25
streamlit
beautifulsoup4
